+++ sigenergy2mqtt/devices/plant/plant.py
@@ -1,5 +1,6 @@
 import asyncio
 import logging
+from datetime import timedelta, timezone
 from typing import cast
 
 import sigenergy2mqtt.sensors.plant_derived as derived
@@ -72,18 +73,20 @@ class PowerPlant(ModbusDevice):
                 logging.info(f"{self.log_identity} ESS Pre-Heating device not registered because pre-heating sensors not found")
 
     async def _register_sensors(self, fw: FirmwareVersion, output_type: int, power_phases: int, modbus_client: ModbusClient) -> None:
+        plant_time_zone = ro.SystemTimeZone(self.plant_index)
         rated_charging_power = ro.PlantRatedChargingPower(self.plant_index)
         rated_discharging_power = ro.PlantRatedDischargingPower(self.plant_index)
         # Fetch async values in parallel
-        rcp_value, rdp_value = await asyncio.gather(
+        tz_offset, rcp_value, rdp_value = await asyncio.gather(
+            tz_offset := plant_time_zone.get_state(modbus_client=modbus_client, raw=True),
             rated_charging_power.get_state(modbus_client=modbus_client),
             rated_discharging_power.get_state(modbus_client=modbus_client),
         )
 
         self.has_battery = isinstance(self._device_type, HybridInverter) and cast(float, rcp_value) > 0.0
 
-        self._add_sensor(ro.SystemTime(self.plant_index))
-        self._add_sensor(ro.SystemTimeZone(self.plant_index))
+        self._add_sensor(ro.SystemTime(self.plant_index, timezone(timedelta(minutes=cast(int, tz_offset)))))
+        self._add_sensor(plant_time_zone)
         self._add_sensor(ro.EMSWorkMode(self.plant_index))
         self._add_sensor(ro.MaxActivePower(self.plant_index))
         self._add_sensor(ro.MaxApparentPower(self.plant_index))